feat(demo5): configurable cycle pause, 32-bit stats counters, stats on by default
Make the inter-measurement pause configurable via DS18B20_CYCLE_PAUSE_US
(default 5s to preserve existing behaviour; demo5 overrides to ~1ms) instead
of the hard-coded 5s PAUSE_5S.  Widen every statistics counter and the
histogram bucket array from uint16_t to uint32_t so a long accumulation window
(no longer overflows past 65535 samples) — this is required now that the stats
period can span thousands of rounds.

demo5 now builds with OW_STATS_ENABLE, STATS_DUMP_INTERVAL=5000 and a 1ms
cycle pause by default.  Parasite power is deliberately left as an explicit
opt-in (EXT="-DPARASITE_POWER=1") because it is bus-hardware dependent, not a
property of the demo.  Document the extra flag in demo5.c.

Add a unit test proving the counters survive >65535 ticks/samples without
wrapping.  Update the test accessor shims and assertions to uint32_t.

+++ tests/mock/ow_stats_test_access.h
@@ -6,10 +6,10 @@
 #include <stdint.h>
 
 const ow_stats_sensor_t* ow_stats_test_get_sensor(uint8_t index);
-uint16_t ow_stats_test_get_histogram(uint8_t bucket);
+uint32_t ow_stats_test_get_histogram(uint8_t bucket);
 uint8_t ow_stats_test_get_sensor_count(void);
-uint16_t ow_stats_test_get_total_cycles(void);
-uint16_t ow_stats_test_get_total_errors(void);
+uint32_t ow_stats_test_get_total_cycles(void);
+uint32_t ow_stats_test_get_total_errors(void);
 uint8_t ow_stats_test_get_dump_phase(void);
 uint8_t ow_stats_test_get_dump_sensor(void);
